refactor: move Indactor map to Atlas
